MicroBooNE cfg changes and additions.
CHANGE from using FrameSplitter to FrameFanout.  This is done because
apparently TbbFlow does not yet have a wrapper for FrameSplitter.
This change SHOULD be innocuous.

ADD a new depo set version of the older per-depo simulation subgraph
construction.  This exposes a new `signal_sets` which is like `signal`
but the edges transfer depo sets instead of single depos.  This is
done to avoid the "Pgrapher slowdown" problem.  We should probably
make this new `signal_sets` be default by naming it `signal`.