Fix compiler warning: moving a temporary object prevents copy elision.
Eg:

../img/src/PointTreeBuilding.cxx:190:35: warning: moving a temporary object prevents copy elision [-Wpessimizing-move]
        auto cnode = root->insert(std::move(std::make_unique<Points::node_t>()));

We should not use std::move() here because the return of make_unique() is already an lvalue.  Likewise in code like:

       cnode->insert(Points(std::move(pcs)));

The Points() constructor returns a temporary which is also an lvalue.

C++ will automaticaly move lvalues to their destination (copy elision).

And for reasons I do not yet understand, std::move() defeats that.